Adding service and  unit test.
