--- Kinematicula.Test/LogicViewing/Services/FitView/SceneInViewFittingTest.cs
@@ -1,12 +1,0 @@
-﻿namespace Kinematicula.Test.LogicViewing.Services.FitView;
-
-using Xunit;
-
-public class SceneInViewFittingTest
-{
-    [Fact]
-    public void Test()
-    {
-
-    }
-}
